refactor: controllers에 있던 Game,Referee를 domains로 이동, models 를 data로 폴더명 변경
