+++ src/controllers/index.js
@@ -1,5 +1,3 @@
-import Game from './Game.js';
 import InputController from './InputController.js';
-import Referee from './Referee.js';
 
-export { Game, InputController, Referee };
+export { InputController };
