+++ src/domains/index.js
@@ -1,0 +1,4 @@
+import Game from './Game.js';
+import Referee from './Referee.js';
+
+export { Game, Referee };
